Segregate StoreHosts controller
