+++ app/api/controllers.go
@@ -138,10 +138,6 @@ func createFlowString(flow alerts.AlertFlow) (string, string) {
 	return source.String(), destination.String()
 }
 
-type blockHostRequest struct {
-	Host string `json:"host" binding:"required"` // Host can be IP or URL
-}
-
 func (api *Api) SendAlertNotification(c *gin.Context) {
 	err := api.AlertsSender.SendLastAlertMessages()
 	if err != nil {
@@ -204,17 +200,4 @@ func parseHostResponse(hosts []host.Host) []HostsResponse {
 		response = append(response, h)
 	}
 	return response
-}
-
-func (api *Api) StoreHosts(c *gin.Context) {
-	err := api.HostsStorage.StoreHosts()
-	if err != nil {
-		fmt.Println(err)
-		c.JSON(500, gin.H{"data": "error"})
-		c.AbortWithStatus(http.StatusInternalServerError)
-		return
-	}
-	c.Header("Access-Control-Allow-Origin", "*") //There is a vuln here, that's only for testing purpose.
-	c.Header("Access-Control-Allow-Methods", "POST")
-	c.JSON(http.StatusOK, gin.H{"message": "ok"})
 }
